requirements.txt
tensorflow-cpu
transformers
streamlit
